% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_DB_response}
\alias{check_DB_response}
\title{check_DB_response}
\usage{
check_DB_response(ret, r, output)
}
\arguments{
\item{ret}{api response}

\item{r}{reactive values}

\item{output}{rshiny output elements}
}
\description{
Checks response from DB query and shows result accordingly
}
