% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{extract_id}
\alias{extract_id}
\title{extract_id}
\usage{
extract_id(input_string)
}
\arguments{
\item{input_string}{character}
}
\value{
Extracted name, name_given and birthdate
}
\description{
Extracts name, given name and birthdate from input. Can then be used to query database
}
