% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_validator}
\alias{get_input_validator}
\title{get_input_validator}
\usage{
get_input_validator(input)
}
\arguments{
\item{input}{rshiny input elements}
}
\value{
TRUE if all input checks have passed, otherwise FALSE
}
\description{
Checks certain fields for correct input
}
