% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latest_entry}
\alias{get_latest_entry}
\title{get_latest_entry}
\usage{
get_latest_entry(data, patientid)
}
\arguments{
\item{data}{tibble}

\item{patientid}{id number}
}
\value{
One row per patient
}
\description{
Selects most current row for given ID
}
