% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_postalcodes_germany}
\alias{get_postalcodes_germany}
\title{Function get_postalcodes_germany}
\usage{
get_postalcodes_germany()
}
\value{
char list
}
\description{
getting list of german postal codes and city names
}
