% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_count_values_bg}
\alias{init_count_values_bg}
\title{Function init_count_values_bg}
\usage{
init_count_values_bg(output, session)
}
\arguments{
\item{output}{output elements}

\item{session}{rshiny session object}
}
\description{
Init background job for summary info text
}
