% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database_ui.R
\name{init_dev_info}
\alias{init_dev_info}
\title{Function}
\usage{
init_dev_info(output)
}
\arguments{
\item{output}{rshiny output elements}
}
\description{
Adding note that on dev/test we only see some patients but not all
}
