% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_reactives.R
\name{init_reactives}
\alias{init_reactives}
\title{Function init_reactives}
\usage{
init_reactives(input, output)
}
\arguments{
\item{input}{input elements}

\item{output}{output elements}
}
\value{
reactive values
}
\description{
defining global reactive values
}
