% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_user_text}
\alias{init_user_text}
\title{Function init_user_text}
\usage{
init_user_text(session, output)
}
\arguments{
\item{session}{shiny session}

\item{output}{shiny elements}
}
\description{
Init user text
}
