% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_input_patientdata}
\alias{load_input_patientdata}
\title{load_input_patientdata}
\usage{
load_input_patientdata(session, ret)
}
\arguments{
\item{session}{rshiny sessions object}

\item{ret}{tibble row}
}
\description{
Replaces input fields in GUI with values from database
}
