% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_aorta_abd}
\alias{observe_aorta_abd}
\title{Function observe_aorta_abd}
\usage{
observe_aorta_abd(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe aorta_abd, i.e. replace input aorta_abd
}
