% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_aorta_asc}
\alias{observe_aorta_asc}
\title{Function observe_aorta_asc}
\usage{
observe_aorta_asc(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe aorta_asc, i.e. replace input aorta_asc
}
