% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_bmi}
\alias{observe_bmi}
\title{Function observe_bmi}
\usage{
observe_bmi(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe BMI, i.e. calculate BMI
}
