% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_comment}
\alias{observe_comment}
\title{Function observe_name}
\usage{
observe_comment(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe comment
}
