% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_dob}
\alias{observe_dob}
\title{Function observe_dob}
\usage{
observe_dob(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe_dob
}
