% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_dwh_updates}
\alias{observe_dwh_updates}
\title{Function observe_dwh_updates}
\usage{
observe_dwh_updates(session, r)
}
\arguments{
\item{session}{shiny sessions}

\item{r}{shiny reactive values}
}
\description{
observe dwh udpates, i.e. update choice selection list
}
