% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database_ui.R
\name{observe_export}
\alias{observe_export}
\title{Function observe_export}
\usage{
observe_export(output, session)
}
\arguments{
\item{output}{shiny output elements}

\item{session}{rhsiny session}
}
\description{
export button handler
}
