% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_first_name}
\alias{observe_first_name}
\title{Function observe_first_name}
\usage{
observe_first_name(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe name
}
