% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_followup_therapy}
\alias{observe_followup_therapy}
\title{Function observe_followup_therapy}
\usage{
observe_followup_therapy(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe followup_therapy
}
