% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_fu}
\alias{observe_fu}
\title{Function observe_fu}
\usage{
observe_fu(input)
}
\arguments{
\item{input}{shiny input elements}
}
\description{
observe follow up fields to make them disable or enable
}
