% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_height}
\alias{observe_height}
\title{Function observe_height}
\usage{
observe_height(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe body height, i.e. replace input height
}
