% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_pre_surgery}
\alias{observe_pre_surgery}
\title{Function observe_pre_surgery}
\usage{
observe_pre_surgery(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe pre_surgery
}
