% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database_ui.R
\name{observe_refresh}
\alias{observe_refresh}
\title{Function observe_refresh}
\usage{
observe_refresh(input, output)
}
\arguments{
\item{input}{shiny input values}

\item{output}{shiny output elements}
}
\description{
refresh table
}
