% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_save}
\alias{observe_save}
\title{Function observe_new}
\usage{
observe_save(session, r, input, output, iv)
}
\arguments{
\item{session}{shiny sessions}

\item{r}{reactive values}

\item{input}{shiny input values}

\item{output}{shiny output values}

\item{iv}{input validator}
}
\description{
observe save button to save to DB
}
