% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_selection}
\alias{observe_selection}
\title{Function observe_selection}
\usage{
observe_selection(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input values}
}
\description{
observe patient selection field
}
