% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_ui.R
\name{observe_weight}
\alias{observe_weight}
\title{Function observe_weight}
\usage{
observe_weight(session, input)
}
\arguments{
\item{session}{shiny sessions}

\item{input}{shiny input elements}
}
\description{
observe body weight, i.e. replace input weight
}
