% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_to_PSA}
\alias{write_to_PSA}
\title{write_to_PSA}
\usage{
write_to_PSA(df, session, r)
}
\arguments{
\item{df}{dataframe}

\item{schema}{character}
}
\value{
.API_response object
}
\description{
Wrapper to write to DWH PSA
}
